+++ fw_al1mb1/libs/usbcfg/usbcfg.h
@@ -14,6 +14,11 @@
     limitations under the License.
 */
 
+/*
+   With compliance of the license:
+   Portions modified from original.
+*/
+
 #ifndef USBCFG_H
 #define USBCFG_H